+++ adapters/Tapjoy/TapjoyAdapter/GADMAdapterTapjoyConstants.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import Foundation;
-@import GoogleMobileAds;
+// Adapter version string
+NSString *const kGADMAdapterTapjoyVersion = @"12.2.0.0";
 
-/// Adapter for communicating with the Tapjoy Network to fetch interstitial ads.
-@interface GADMAdapterTapjoy : NSObject <GADMAdNetworkAdapter>
+// Adapter mediation agent
+NSString *const kGADMAdapterTapjoyMediationAgent = @"admob";
 
-@end
+// Tapjoy SDK key
+NSString *const kGADMAdapterTapjoySdkKey = @"sdkKey";
+
+// Tapjoy placemnt name
+NSString *const kGADMAdapterTapjoyPlacementKey = @"placementName";
+
+NSString *const kGADMAdapterTapjoyErrorDomain = @"com.google.mediation.tapjoy";