+++ adapters/Tapjoy/TapjoyAdapter/GADMediationAdapterTapjoy.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import Foundation;
+#import <Foundation/Foundation.h>
 @import GoogleMobileAds;
 
-/// Adapter for communicating with the Tapjoy Network to fetch interstitial ads.
-@interface GADMAdapterTapjoy : NSObject <GADMAdNetworkAdapter>
+NS_ASSUME_NONNULL_BEGIN
+
+@interface GADMediationAdapterTapjoy : NSObject <GADMediationAdapter>
 
 @end
+
+NS_ASSUME_NONNULL_END